Rename getDatasetHelpers to utils/prefix
These prefix-handling functions are used more widely than the Charon
getDataset endpoint handler.
