+++ src/auspicePaths.js
@@ -1,7 +1,7 @@
 const fetch = require("node-fetch");
 const sources = require("./sources");
 const { warn } = require("./utils");
-const { splitPrefixIntoParts, parsePrefix } = require("./getDatasetHelpers");
+const { splitPrefixIntoParts, parsePrefix } = require("./utils/prefix");
 
 /**
  * Auspice core builds have URLs where the first (`/`-separated) field matches the following list.
