+++ src/utils/prefix.js
@@ -1,7 +1,7 @@
 const {NotFound} = require("http-errors");
 
-const utils = require("./utils");
-const sources = require("./sources");
+const utils = require(".");
+const sources = require("../sources");
 
 /* All of the logic for mapping a dataset or narratives URL ("prefix") to a
  * source + path is intentionally encapsulated contained here.  This function
